output/42/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyrag"
version = "0.1.0"
description = "Docling-powered modular RAG CLI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "SHAI", email = "ops@shai.dev" },
]
dependencies = [
    "langchain==1.0.8",
    "langchain-core==1.0.6",
    "langchain-community==0.4.1",
    "langchain-text-splitters==1.0.0",
    "langchain-huggingface==1.0.1",
    "langchain-milvus==0.3.0",
    "docling==2.4.0",
    "sentence-transformers==3.0.1",
    "torch==2.4.1",
    "onnxruntime==1.18.0",
    "pymilvus==2.6.0",
    "milvus-lite==2.4.7",
    "python-dotenv==1.0.1",
    "typer==0.12.5",
    "rich==13.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-mock==3.14.0",
    "ruff==0.6.3",
]

[project.urls]
Source = "https://github.com/shai/shai"

[project.scripts]
pyrag = "pyrag.__main__:main"

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
